requirements.txt
Django==2.1.3
django-extensions==1.9.7
django-stdimage==3.0.0
olefile==0.44
Pillow==5.2.0
progressbar2==3.34.3
python-utils==2.2.0
pytz==2017.3
six==1.11.0
typing==3.6.2
django-robots==3.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
idna==2.8
oauthlib==2.1.0
PyJWT==1.7.1
python3-openid==3.1.0
requests==2.21.0
requests-oauthlib==1.0.0
urllib3==1.24.1
django-widget-tweaks==1.4.3